Expose the notification log
Google added a way to see your past notifications but is buried. The stock
hoes were accessing this by way of the Settings shortcuts but I feel like
is time for our users to see this without having to do any fuckery.

In case you don't know what WTF this is...

androidpit.com/how-to-recover-lost-notifications-on-android

Change-Id: I735b3d6c0a94f3d5bed867a2988056f6545df1df
Signed-off-by: SagarMakhar <sagarmakhar@gmail.com>
Signed-off-by: RakeshBatra <raakesh.batra@rediffmail.com>